INTERNAL: Creating piped collection operation future.
